+++ UnitTests/PaymentButtonsTests/PayPalCreditButton_Tests.swift
@@ -14,7 +14,7 @@ class PayPalCreditButton_Tests: XCTestCase {
         let sut = PayPalCreditButton()
         XCTAssertEqual(sut.edges, PaymentButtonEdges.softEdges)
         XCTAssertEqual(sut.size, PaymentButtonSize.collapsed)
-        XCTAssertEqual(sut.color, PaymentButtonColor.white)
+        XCTAssertEqual(sut.color, PaymentButtonColor.darkBlue)
         XCTAssertNil(sut.insets)
         XCTAssertNil(sut.label)
     }
